# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/routes/{routeId}/places": {
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Create a new place for a specific route",
                "operationId": "97675972579a22aa5e7a9fd3705f2a46",
                "parameters": [
                    {
                        "name": "routeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "/schemas/PlaceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Place created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/Place"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                }
            }
        },
...
        "/api/routes/{routeId}/places/{placeId}": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Get a specific place",
                "operationId": "578ec6753b1d1cbb88e0c1c700a695a9",
                "parameters": [
                    {
                        "name": "routeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Place retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/Place"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Place does not belong to this route."
                    }
                }
            },
            "put": {
                "tags": [
                    "Places"
                ],
                "summary": "Update a specific place",
                "operationId": "8104ad1e06e6d72f94560f09c9a4ed76",
                "parameters": [
                    {
                        "name": "routeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "/schemas/PlaceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Place updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/Place"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You are not authorized to update this place."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Places"
                ],
                "summary": "Delete a specific place",
                "operationId": "05c9c928c324b25d45861a9f03713b22",
                "parameters": [
                    {
                        "name": "routeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Place deleted successfully."
                    },
                    "403": {
                        "description": "You are not authorized to delete this place."
                    }
                }
            }
        },
        "/weather/{route}/{place}": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Get weather data for a specific place",
                "operationId": "1895d509aae47a9054af313c2aaf604b",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "place",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with weather data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/WeatherResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Place does not belong to this route",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Place does not belong to this route"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch weather data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch weather data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hotel/{route}/{place}": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get hotel data for a specific place",
                "operationId": "5dfd092be670b89f99ee33d0c776663c",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "place",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with hotel data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/schemas/HotelResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Place does not belong to this route",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Place does not belong to this route"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch hotels data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch hotels data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}